% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{DataCompositionContainer}
\alias{DataCompositionContainer}
\title{DataCompositionContainer Class}
\description{
A class container that holds the long and wide catch-at-length and mean-weight-at-length dataframes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataCompositionContainer-new}{\code{DataCompositionContainer$new()}}
\item \href{#method-DataCompositionContainer-clone}{\code{DataCompositionContainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCompositionContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataCompositionContainer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCompositionContainer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCompositionContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataCompositionContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCompositionContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
