% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{DataCompositionContext}
\alias{DataCompositionContext}
\title{SpeciesGearDataContext Class}
\description{
A class that provide context about
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataCompositionContext-new}{\code{DataCompositionContext$new()}}
\item \href{#method-DataCompositionContext-clone}{\code{DataCompositionContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCompositionContext-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataCompositionContext-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCompositionContext$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataCompositionContext-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataCompositionContext-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCompositionContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
