% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MixinUtilities}
\alias{MixinUtilities}
\title{Mixin Utility class}
\description{
It holds a few util methods used across all classes in the this R package
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MixinUtilities-create_empty_dataframe}{\code{MixinUtilities$create_empty_dataframe()}}
\item \href{#method-MixinUtilities-clone}{\code{MixinUtilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-create_empty_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-create_empty_dataframe}{}}}
\subsection{Method \code{create_empty_dataframe()}}{
Create an empty dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$create_empty_dataframe(col_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col_names}}{vector with the column names of the empty dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
