% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utpb_data_composition.R
\name{UtpbDataComposition}
\alias{UtpbDataComposition}
\title{UtpbDataComposition  Class}
\description{
Custom class to build the catches and mean-weight matrcies for the UTPB database
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{a long dataframe with the length and weight data}

\item{\code{time_col}}{name of the column that holds dates}

\item{\code{size_col}}{name of the column that holds lengths}

\item{\code{weight_col}}{name of the column that holds weights}

\item{\code{mean_weight_col}}{name of the column that will hold the meanweights}

\item{\code{interval_col}}{name of the column that will hold the size interval (as opposed to midpoints)}

\item{\code{midpoint_col}}{name of the column that will hold the length classes}

\item{\code{freq_col}}{name of the column that will hold the the length totals per class}

\item{\code{composition_results}}{name of the column that will hold the the length totals per class}

\item{\code{linf}}{asymptotic length}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtpbDataComposition-new}{\code{UtpbDataComposition$new()}}
\item \href{#method-UtpbDataComposition-build_catch_at_length}{\code{UtpbDataComposition$build_catch_at_length()}}
\item \href{#method-UtpbDataComposition-build_talla_and_weight_composition_matrices}{\code{UtpbDataComposition$build_talla_and_weight_composition_matrices()}}
\item \href{#method-UtpbDataComposition-clone}{\code{UtpbDataComposition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDataComposition-new"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDataComposition-new}{}}}
\subsection{Method \code{new()}}{
Initialise the SpeciesDataComposition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDataComposition$new(data, sp_arte_context)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a long dataframe with the length and weight data about a single species}

\item{\code{sp_arte_context}}{a DataCompositionContext class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDataComposition-build_catch_at_length"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDataComposition-build_catch_at_length}{}}}
\subsection{Method \code{build_catch_at_length()}}{
Build a year-basis catch-at-length composition matrix when the weight column is
not debugged, therefore it could be NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDataComposition$build_catch_at_length(
  bindwidth,
  col_prefix,
  min_padding,
  up_to_linf = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bindwidth}}{size of each length class}

\item{\code{col_prefix}}{prefix added to the each column representing a year}

\item{\code{min_padding}}{padding that is substracted to the minimum length class}

\item{\code{up_to_linf}}{boolean flag indicating whether the length classed should be capped}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDataComposition-build_talla_and_weight_composition_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDataComposition-build_talla_and_weight_composition_matrices}{}}}
\subsection{Method \code{build_talla_and_weight_composition_matrices()}}{
Build a year-basis catch-at-length and mean-weight composition matrices both in long and wide format
(4 matrices, 2 for catch_at_length and 2 for mean-weight)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDataComposition$build_talla_and_weight_composition_matrices(
  bindwidth,
  col_prefix,
  min_padding,
  up_to_linf = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bindwidth}}{size of each length class}

\item{\code{col_prefix}}{prefix added to the each column representing a year}

\item{\code{min_padding}}{padding that is substracted to the minimum length class}

\item{\code{up_to_linf}}{boolean flag indicating whether the length classed should be capped}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDataComposition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDataComposition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDataComposition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
